requirements.txt
autopep8==2.3.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pycodestyle==2.12.1
PyJWT==2.5.0
Werkzeug==3.1.3
